refactor(action): reorganize multiplayer session related actions
