+++ apps/next-app/src/lib/safe-action/session/multiplayer/use-create-room.ts
@@ -3,7 +3,7 @@ import { useAction } from "next-safe-action/hooks"
 import { toast } from "sonner"
 
 // actions
-import { createRoom } from "@/server/action/room-action"
+import { createRoom } from "@/server/action/session/multiplayer-action"
 
 // utils
 import { handleServerError } from "@/lib/util/error"
@@ -12,8 +12,12 @@ export const useCreateRoomAction = () => {
   const router = useRouter()
 
   return useAction(createRoom, {
-    onExecute() { toast.loading("Creating room...", { id: "room:create" }) },
-    onSettled() { toast.dismiss("room:create") },
+    onExecute() {
+      toast.loading("Creating room...", { id: "room:create" })
+    },
+    onSettled() {
+      toast.dismiss("room:create")
+    },
     onSuccess() {
       toast.dismiss("room:create:error")
 
@@ -23,10 +27,7 @@ export const useCreateRoomAction = () => {
       })
     },
     onError({ error }) {
-      if (
-        error.serverError?.key === "ACTIVE_SESSION" ||
-        error.serverError?.key === "ACTIVE_ROOM"
-      ) {
+      if (error.serverError?.key === "ACTIVE_SESSION" || error.serverError?.key === "ACTIVE_ROOM") {
         const { message, description } = error.serverError
 
         toast.warning(message, {
@@ -35,7 +36,9 @@ export const useCreateRoomAction = () => {
           duration: 10000,
           action: {
             label: "Reconnect",
-            onClick() { router.push("/game/multiplayer") }
+            onClick() {
+              router.push("/game/multiplayer")
+            }
           }
         })
         return
